async-task-service/requirements.txt
redis>=5.0,<6
redis[hiredis]>=5.0,<6
python-dotenv>=1.0,<2
httpx>=0.24,<1
requests==2.32.5
